// sdk/client/client-node-sdk/src/node-sdk/model/badGatewayErrorMessageResponse.ts
/**
 * Lead Scraping Service API
 * Vector Lead Scraping Service API - Manages Lead Scraping Jobs
 *
 * The version of the OpenAPI document: 1.0
 * Contact: yoanyomba@vector.ai
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { ErrorResponse } from './errorResponse';

/**
* Represents errors when the server received an invalid response from an upstream server
*/
export class BadGatewayErrorMessageResponse {
    'code'?: number;
    'message'?: string;
    'upstreamService'?: string;
    'errorResponse'?: ErrorResponse;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "code",
            "baseName": "code",
            "type": "number"
        },
        {
            "name": "message",
            "baseName": "message",
            "type": "string"
        },
        {
            "name": "upstreamService",
            "baseName": "upstreamService",
            "type": "string"
        },
        {
            "name": "errorResponse",
            "baseName": "errorResponse",
            "type": "ErrorResponse"
        }    ];

    static getAttributeTypeMap() {
        return BadGatewayErrorMessageResponse.attributeTypeMap;
    }
}